requirements.txt
google-api-python-client>=2.170.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
Pillow>=10.4.0
openai>=1.82.0
